Fix overflow issue in letterboxing kernel (#2)
When we are using 1640x1232 camera resolution, our height padding ends
up rounding to 79 instead of 80. This results in an buffer overrun -
illegal memory access in the CUDA kernel when accessing the input image.

The fix is to correctly round the paddings so we are staying in the
bounds of the input image.

The memory error was not reported on previous JetPacks (<6.2) due to
using a legacy CUDA version.